backend/requirements.txt
PyMuPDF==1.23.14
faiss-cpu==1.7.4
sentence-transformers==2.2.2
Pillow==10.0.1
pytesseract==0.3.10
numpy==1.24.3
torch==2.0.1
transformers==4.33.2